Change XML file output location
Output the final XML file for updates to the root directory and then
manually upload the changed file to the server. This way most users
won't download the patch until all the files are uploaded to the server
since I can wait and then upload the final XML file after they are
uploaded.